+++ AssetsFlow.Client.React/CHANGELOG.md
@@ -5,6 +5,7 @@
 fix: cotent edit popup open> disable tooltip
 allow edit on 1/2 overview
 fix: close trade not working
+fix: close trade not refreshing - parse new-reduce value
 
 ### Fixed