+++ AssetsFlow.Server/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [0.9.225-b] - 2025-08-...
-
+- validate net amount on close trade, return reduce element on close
 
 ## [0.9.224-b] - 2025-08-11
 - More .env work